requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
beautifulsoup4==4.13.4
certifi==2026.2.25
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
